Update 230706Minutes.md
Hi Martino,

Thank you so much for sharing the minutes of the meeting and the follow-up message. Much appreciated.

I have added a few edits. Only one concern about what you call "curriculum-level choices". I have checked my notes and rephrased a couple of points accordingly. My understanding is that the LPGC will be responsible for making decisions on the addition/replacement/retirement of lessons, but agreement with the CAC and consultation with the corresponding maintainers is required for the LPGC to proceed with coordinating such changes. Please check.

Lastly, did not we agree this CAC-LPGC relationship will be added to the existing rubric, once is approved by the CAC?

Thanks heaps. See you on the 10th.

Cheers,
Valentina